Refactor file structure and strengthen prop types
The code refactor includes moving `ContentResultBody.tsx` and `ContentResultHeader.tsx` to a new `ContentResult` subdirectory, adjusting import paths accordingly. Additionally, the types of the component props have been explicitly set to `Readonly` to prevent potential mutation. This protects the integrity of data passed down and enhances code maintainability.
