+++ frontend/src/components/ContentResult/ContentResultBody.tsx
@@ -1,12 +1,14 @@
 import { Flex } from "@chakra-ui/react";
-import ContentCard from "./ContentCard/ContentCard.tsx";
-import { contentType } from "../model/contentModel.ts";
+import ContentCard from "../ContentCard/ContentCard.tsx";
+import { contentType } from "../../model/contentModel.ts";
 
 type ContentResultBodyProps = {
   content: contentType[];
 };
 
-export default function ContentResultBody({ content }: ContentResultBodyProps) {
+export default function ContentResultBody({
+  content,
+}: Readonly<ContentResultBodyProps>) {
   return (
     <Flex flexDirection="column" gap={4}>
       {content.map((content) => {
