+++ frontend/src/components/ContentResult/ContentResultHeader.tsx
@@ -1,5 +1,5 @@
 import { Card, Skeleton, Text } from "@chakra-ui/react";
-import { infoType } from "../model/contentModel.ts";
+import { infoType } from "../../model/contentModel.ts";
 
 type ContentFilterResultHeaderProps = {
   info: infoType | undefined;
@@ -11,7 +11,7 @@ function pluralizeResults(count: number) {
 
 export default function ContentResultHeader({
   info,
-}: ContentFilterResultHeaderProps) {
+}: Readonly<ContentFilterResultHeaderProps>) {
   if (info === undefined) {
     return (
       <Card bgColor="gray.100" mb={4} fontSize="sm" p={2}>
